refactor: split ApiClient with a base class
