+++ src/constants.ts
@@ -1,0 +1,11 @@
+export const AUTH_VERSION = '0.1';
+export const APPLICATION_VERSION = '1.0';
+export const CONFD_VERSION = '1.1';
+export const CTIDNG_VERSION = '1.0';
+export const DIRD_VERSION = '0.1';
+export const CALL_LOGD_VERSION = '1.0';
+export const CHATD_VERSION = '1.0';
+export const CALLD_VERSION = '1.0';
+export const AGENTD_VERSION = '1.0';
+export const WEBHOOKD_VERSION = '1.0';
+export const AMID_VERSION = '1.0';
